requirements.txt
numpy
pandas
seaborn
matplotlib
gdown==4.6.0
scipy
scikit-learn
ensure
python-box
pyYAML
-e .
